Add jackson and joda-time as dependency to emr-dynamodb-hadoop
in EMR we always have /usr/share/aws/aws-java-sdk/*.jar including
/usr/share/aws/aws-java-sdk/aws-java-sdk-core-*.jar which didn't shaded
those jackson library as part of the import line.

As such that causes the conflict when both aws-java-sdk-bundle and
aws-java-sdk-core in the classpath, always we loaded aws-java-sdk-core
first. and aws-java-sdk-core were calling jackson directly instead of
a shaded jackson, then NoSuchMethodError will be issued.